Updated Scenes
Added a d6 scene to be a fully functional d6 die node, containing a DieNode child node and a Sides child node with one Direction node for each face.
Created a PhysicsMaterial instance, dice-physics, to ensure all dice bounce with the same intensity.

Added the UI scene to the Main scene. Added a d6 in the Main Scene, ready to fall and tumble.