Updated UI
Updated layout of DiceButtons and RollButtons so they will move appropriately (or rather, stay where they should be) when the window resizes.
Rearranged both sets of buttons to be wholly vertical.
Updated Mouse properties on the DiceButtons buttons to display a pointing hand.
Updated the Button Mask on the DiceButtons to accept Mouse Right as well in preparation for right-click to subtract implementation.