Refined DiceManager
Added new Properties for each of the random generation limits, as well as SpawnHeight and GroupImpulse. SpawnHeight is self-explanatory but GroupImpulse is a boolean determining whether or not a single impulse vector is shared by each die type or if a new impulse vector is generated for every individual die.

Cleaned up the resource loading code by using a godot::String and formatting in the number of sides.