Updated UILogic
Normalizing custom signal names to PascalCase.
Changed pressHandler implementation for the dice buttons to connect to the gui_input signal and, on right click, decrease the requested quantity of that die type or, on left click, increase the requested quantity of that die type.
Added new signal, "UpdateEquation," for use by the forthcoming UI control that will display the current requested equation to the user.