Refining Dice Rolling
Added function getDirection to Direction node, returning the non-normalized global rotation vector; Updated getDirectionUnit to not round the return value.

Updated DieNode::getValue() to be more precise; When a direction matches Vector3.UP precisely, use that value; Otherwise, use the dot product of the non-normalized direction and Vector3.UP and find the largest dot product, using that Direction's value. It resolves issues of dice getting stuck on an edge but it causes some inconsistencies with the d20 given how similar the angles are for each face.