Updated Nodes
Added DisplayValueLabel and GuessNotFlat properties to DieNode, controlling whether the Label3D is updated/displayed and how fuzzy the algorithm in GetValue should be, respectively.

Added the same properties to DiceManager, as that is the node that actually generates the dice. Applies the editor configured values of these two properties to each DieNode it instantiates.