Updated UILogic Node
Added the wiring to toggle what values in the roll's results will or will not be displayed.
Added a boolean value to allow rolls to be repeated without needing to rebuild the equation; The dice count map is cleared when rolledSinceCountUpdate is true and the user clicks one of the dice buttons.